Radar: move utility functions in another file
